refactor(tasks): extract today task filter
